plan: switch to simpler baseline master scan table definition and format

+++ plan/src/plan_list.c
@@ -119,10 +119,10 @@ int plan_list_write (FILE *fp, double (*mirror_tilt)(double),
              if (0 != mkjdtimestr (tstart_jd, buf, sizeof(buf)))
                return -1;
 
-             if (fprintf (fp, "%0.16e,%0.3f,%d,%d,%0.3f,\"%s\"\n",
+             if (fprintf (fp, "%0.16e,%0.3f,%0.1f,%d,%0.3f,\"%s\"\n",
                           tstart_tai,
                           entry->scan_duration,
-                          (int) fsw_xstart,
+                          fsw_xstart,
                           entry->num_steps,
                           entry->integration_time,
                           buf) < 0)
